module exported via discover method
