+++ templates/add-action/generic/templates/stub-action.e2e.js
