+++ templates/add-action/generic/templates/stub-action.js
