+++ templates/add-web-assets/exc-react/index.js
