+++ templates/add-web-assets/exc-react/templates/index.html
@@ -19,8 +19,8 @@ governing permissions and limitations under the License.
   <title><%= projectName %></title>
 </head>
 <body>
-  <noscript>You need to enable JavaScript to run this app.</noscript>
-  <div id="root"></div>
-  <script src="./src/index.js" async="true"></script>
+<noscript>You need to enable JavaScript to run this app.</noscript>
+<div id="root"></div>
+<script src="./src/index.js" async="true"></script>
 </body>
 </html>
