+++ templates/add-web-assets/exc-react/templates/src/components/About.js
