+++ templates/add-web-assets/exc-react/templates/src/components/App.js
