+++ templates/add-web-assets/exc-react/templates/src/components/Home.js
