+++ templates/add-web-assets/exc-react/templates/src/components/SideBar.js
