+++ templates/add-web-assets/exc-react/templates/src/config.json
