+++ templates/add-web-assets/exc-react/templates/src/index.js
